GammaCC.tcl is a general-purpose shared-object generator that will eventually compile Gamma code.

+++ ramspice_types.h
@@ -15,7 +15,7 @@ int verbose;
     typedef float half_scalar;
 typedef union {
     ordinal o;
-    double s;
+    float s;
     unsigned char b[BYTES_PER_SCALAR];
     half_scalar h[2];
     void *v;
